get ready for release, before update and build

+++ app/src/main/java/com/example/photo_post/models/Instrument.kt
@@ -1,7 +1,7 @@
 package com.example.photo_post.models
 
 data class Instrument(
-    val instrId: Int,
+    val instrId: Long,
     val instrName: String,
     val instrQr: String,
     val instrProps: String
