+++ app/src/main/java/com/example/photo_post/models/JsonModel.kt
@@ -1,12 +1,9 @@
 package com.example.photo_post.models
 
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-
 data class JsonModel(
-    val cart_id: Int,
+    val cart_id: Long,
     val cart_name: String,
-    val instr_id: Int,
+    val instr_id: Long,
     val cart_user_pass: String,
     val quantity: Double,
     val instr_name: String,
